Learned about time zones.
I learned about how you can set a new DateTime set to a time zone and that it is important that you store all datetime values in a database with UTC and when you pull a value and display it to a user. You then need to convert it to their timezone.